refactor: split gateway admin modules
